ticket:4521: 64bit align fix on branch builds & passes all make check tests
git-svn-id: http://source.icu-project.org/repos/icu/icu@27645 251d0590-4201-4cf1-90de-194747b24ca1

+++ branches/pedberg/fix4521/source/common/uvectr64.cpp
@@ -124,7 +124,7 @@ UBool UVector64::expandCapacity(int32_t minimumCapacity, UErrorCode &status) {
     if (maxCapacity > 0 && newCap > maxCapacity) {
         newCap = maxCapacity;
     }
-    int32_t* newElems = (int64_t *)uprv_realloc(elements, sizeof(int64_t)*newCap);
+    int64_t* newElems = (int64_t *)uprv_realloc(elements, sizeof(int64_t)*newCap);
     if (newElems == NULL) {
         // We keep the original contents on the memory failure on realloc.
         status = U_MEMORY_ALLOCATION_ERROR;
